+++ CHANGELOG.md
@@ -11,7 +11,7 @@
  - Adding shortcut for loading Sprites and SpriteAnimation from the global cache
  - Adding loading methods for the different `ParallaxComponent` parts and refactor how the delta velocity works
  - Add tests for `Timer` and fix a bug where `progress` was not reported correctly
- - Refactored the `SpriteBatch` class to be more elegant.
+ - Refactored the `SpriteBatch` class to be more elegant and to support `Vector2`.
  - Added fallback support for the web on the `SpriteBatch` class
  - Added missing documentation on the `SpriteBatch` class
  - Added an utility method to load a `SpriteBatch` on the `Game` class